gate_draw flowSet Method

Manually draw gates around populations for analysis of flow cytometry data.

# S4 method for flowSet
gate_draw(x, select = NULL, channels = NULL,
  alias = NULL, display = 1/length(x), type = "polygon",
  axis = "x", density_smooth = 1.5, label = TRUE, plot = TRUE, ...)

Arguments

x
object of class flowSet.

select
vector containing the indicies of samples within gs to use for plotting.

channels
vector of channel names to use for plotting, can be of length 1 for 1-D density histogram or length 2 for 2-D scatter plot.

alias
the name(s) of the populations to be gated. If multiple population names are supplied (e.g. c("CD3,"CD4)) multiple gates will be returned. alias is NULL by default which will halt the gating routine.

display
numeric [0,1] to control the percentage of events to be plotted. Specifying a value for display can substantial improve plotting speed for less powerful machines.

type
vector of gate type names used to construct the gates. Multiple gate types are supported but should be accompanied with an alias argument of the same length (i.e. one type per alias). Supported gate types are polygon, rectangle, ellipse, threshold, boundary, interval, quadrant and web which can be abbreviated as upper or lower case first letters as well. Default type is "interval" for 1D gates and "polygon" for 2D gates.

axis
indicates whether the "x" or "y" axis should be gated for 2-D interval gates.

density_smooth
smoothing factor passed to density for 1-D plots (defaults to 1.5).

label
logical indicating whether to include cyto_plot_label for the gated population(s), TRUE by default.

plot
logical indicating whether a plot should be drawn, set to TRUE by default.

...
additional arguments for cyto_plot,flowSet-method.

Value

a filters list containing the drawn gate objects.

See also

Examples

# NOT RUN {
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# draw gates using gate_draw - add contour lines & overlay control
gate_draw(fs,
  channels = c("FSC-A", "SSC-A"),
  alias = "Cells",
  type = "polygon",
  contour_lines = 15,
  overlay = fs[[1]]
)

# gt is a filters object containing the contructed polygonGate
gt[[1]]
# }