Draw Polygon Gate(s) Around Populations.

gate_polygon_draw constructs an interactive plotting window to allow manual selection of the co-ordinates of a polygon gate(s) (through mouse click) which are constructed into polygonGate objects and stored in a filters list.

gate_polygon_draw(fr, alias = NULL, channels, plot = TRUE,
  label = TRUE, ...)

Arguments

fr
a flowFrame object containing the flow cytometry data for plotting and gating.

alias
the name(s) of the populations to be gated. If multiple population names are supplied (e.g. c("CD3","CD4")) multiple gates will be returned. alias is NULL by default which will halt the gating routine.

channels
vector of channel names to use for plotting, can be of length 1 for 1-D density histogram or length 2 for 2-D scatter plot.

plot
logical indicating whether the data should be plotted. This feature allows for constructing gates of different types over existing plots which may already contain a different gate type.

label
logical indicating whether to include cyto_plot_label for the gated population(s), TRUE by default.

...
additional arguments for cyto_plot,flowFrame-method.

Value

afilters list containing the constructed polygonGate object(s).

See also

Examples

# NOT RUN {
# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get polygonGate using gate_polygon_draw
pg <- gate_polygon_draw(fs[[4]],
  alias = "Cells",
  channels = c("FSC-A", "SSC-A")
)

# pg is a filters object - extract polygonGate using `[[`
pg[[1]]
# }